+++ packages/ai/src/agent/index.ts
@@ -25,3 +25,4 @@ export {
   type InferAgentUIMessage,
 } from './infer-agent-ui-message';
 export { createAgentUIStreamResponse } from './create-agent-ui-stream-response';
+export { createAgentUIStream } from './create-agent-ui-stream';
